Stage3Wave5の下半分の右側が少し早く出現する問題を解決しました。 原因としては同様にINIT_ARG()関数の戻り値がdoubleではなくintだったためです。 Stage3Wave5について、少しバランス調整も行いました。

+++ program/Character/EnemyCharacter/ZakoCharacter/ZkChrStg3Wv5BLR.cpp
@@ -38,7 +38,7 @@ int ZkChrStg3Wv5BLR::INIT_POS_X(Stg3WAVE5BLR given_lr) {
 	}
 }
 
-int ZkChrStg3Wv5BLR::INIT_ARG(Stg3WAVE5BLR given_lr) {
+double ZkChrStg3Wv5BLR::INIT_ARG(Stg3WAVE5BLR given_lr) {
 	if (given_lr == Stg3WAVE5BLR::LEFT) {
 		return INIT_ARG_L;
 	}
